% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indic_cell.R
\name{Calc_indic_proba}
\alias{Calc_indic_proba}
\title{Title: Calc_indic_proba}
\usage{
Calc_indic_proba(x = reseau_cell[[1]], bin_threshold = 1)
}
\arguments{
\item{x}{Description of the first parameter.}

\item{y}{Description of the second parameter (if applicable).}

\item{...}{Other optional parameters passed to methods.}
}
\value{
Description of the object that the function returns.
If the function doesn't return anything meaningful, you can say `NULL`.
}
\description{
Title: Calc_indic_proba
}
