% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df_depth_env}
\alias{df_depth_env}
\title{Depth range and water position of marine fish species}
\format{
A data frame with 11786 rows and 4 columns:
\describe{
\item{Species}{Scientific name.}
\item{Depth_min, Depth_max}{Depth range (m).}
\item{Env_2}{Position in the water column (FishBase categories).}
}
}
\usage{
df_depth_env
}
\description{
Depth range and water position of marine fish species
}
\keyword{datasets}
