% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_metaweb.R
\name{Distri_correction}
\alias{Distri_correction}
\title{Title: Distri_correction}
\usage{
Distri_correction(data, Lniche, mc.cores = 4)
}
\arguments{
\item{x}{Description of the first parameter.}

\item{y}{Description of the second parameter (if applicable).}

\item{...}{Other optional parameters passed to methods.}
}
\value{
Description of the object that the function returns.
If the function doesn't return anything meaningful, you can say `NULL`.
}
\description{
Title: Distri_correction
}
\details{
Correction for water column distribution
}
