% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indic_cell.R
\name{get_indic_cells}
\alias{get_indic_cells}
\title{Title: metaweb_mod_parameters}
\usage{
get_indic_cells(P_A_data, Lniche, mc.cores = 1)
}
\arguments{
\item{x}{Description of the first parameter.}

\item{y}{Description of the second parameter (if applicable).}

\item{...}{Other optional parameters passed to methods.}
}
\value{
Description of the object that the function returns.
If the function doesn't return anything meaningful, you can say `NULL`.
}
\description{
Title: metaweb_mod_parameters
}
\details{
Find the parameter for the metaweb
}
