% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_traits.R
\name{get_traits_higher_edna}
\alias{get_traits_higher_edna}
\title{Title: get_traits}
\usage{
get_traits_higher_edna(
  data_presence,
  column_species = "species",
  column_taxon = "taxon"
)
}
\arguments{
\item{x}{Description of the first parameter.}

\item{y}{Description of the second parameter (if applicable).}

\item{...}{Other optional parameters passed to methods.}
}
\value{
Description of the object that the function returns.
If the function doesn't return anything meaningful, you can say `NULL`.
}
\description{
# End goal:
}
\details{
Find the parameter for the metaweb - for both species and genus-level here
}
