% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{model}
\alias{model}
\title{Title: model}
\usage{
model(pars, data)
}
\arguments{
\item{x}{Description of the first parameter.}

\item{y}{Description of the second parameter (if applicable).}

\item{...}{Other optional parameters passed to methods.}
}
\value{
Description of the object that the function returns.
If the function doesn't return anything meaningful, you can say `NULL`.
}
\description{
Title: model
}
\details{
Helper
}
