% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metaweb.R
\name{pLMFitted}
\alias{pLMFitted}
\title{Title: pLMFitted}
\usage{
pLMFitted(MPrey, MPred, Pars)
}
\value{
Description of the object that the function returns.
If the function doesn't return anything meaningful, you can say `NULL`.
}
\description{
Title: pLMFitted
}
\details{
Helper for above
}
