% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_metaweb.R
\name{apply_model_metaweb}
\alias{apply_model_metaweb}
\title{Title: apply_model_metaweb}
\usage{
apply_model_metaweb(
  data_traits,
  path_pars,
  length_column = "CommonLengthEstim"
)
}
\arguments{
\item{data_traits}{Dataframe containing trait data (columms...)}

\item{path_pars}{Path to parameters dataframe}
}
\value{
Description of the object that the function returns.
If the function doesn't return anything meaningful, you can say `NULL`.
}
\description{
Clean trait data. Get common lenght if missing (0.6*TL)
}
\details{
Create the metaweb
}
