% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{clean_traits}
\alias{clean_traits}
\title{Title: clean_traits}
\usage{
clean_traits(data_traits, data_presence = NULL, TL_clean = FALSE)
}
\arguments{
\item{x}{Description of the first parameter.}

\item{y}{Description of the second parameter (if applicable).}

\item{...}{Other optional parameters passed to methods.}
}
\value{
Description of the object that the function returns.
If the function doesn't return anything meaningful, you can say `NULL`.
}
\description{
Clean trait data. Get common lenght if missing (0.6*TL)
}
\details{
Clean trait data
}
