% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_metaweb.R
\name{correct_metaweb_fish}
\alias{correct_metaweb_fish}
\title{Title: correct_metaweb_fish}
\usage{
correct_metaweb_fish(
  data_meta,
  data_traits,
  herbivory = TRUE,
  water_pos = TRUE,
  small_fish = TRUE,
  PP_SP_add = TRUE,
  exception_BigFish = TRUE
)
}
\arguments{
\item{x}{Description of the first parameter.}

\item{y}{Description of the second parameter (if applicable).}

\item{...}{Other optional parameters passed to methods.}
}
\value{
Description of the object that the function returns.
If the function doesn't return anything meaningful, you can say `NULL`.
}
\description{
Title: correct_metaweb_fish
}
\details{
Correct metaweb based on general ecology
}
