% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proba.R
\name{correct_metaweb_proba}
\alias{correct_metaweb_proba}
\title{Title: correct_metaweb_proba}
\usage{
correct_metaweb_proba(MW, df_traits)
}
\arguments{
\item{MW}{Adjency matrix (or metaweb)}

\item{y}{Description of the second parameter (if applicable).}

\item{...}{Other optional parameters passed to methods.}
}
\value{
Description of the object that the function returns.
If the function doesn't return anything meaningful, you can say `NULL`.
}
\description{
Title: correct_metaweb_proba
}
\details{
correct the metaweb based on the proba
}
