% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{filter_adjacency}
\alias{filter_adjacency}
\title{Title: filter_adjacency}
\usage{
filter_adjacency(adj_matrix, keep_vec)
}
\arguments{
\item{x}{Description of the first parameter.}

\item{y}{Description of the second parameter (if applicable).}

\item{...}{Other optional parameters passed to methods.}
}
\value{
Description of the object that the function returns.
If the function doesn't return anything meaningful, you can say `NULL`.
}
\description{
Title: filter_adjacency
}
\details{
Helper function for the plot_network
}
