% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{get_fig_eval}
\alias{get_fig_eval}
\title{Title: get_fig_eval}
\usage{
get_fig_eval(data = data_end, target_sp = "Zeus faber", pars = pars)
}
\arguments{
\item{x}{Description of the first parameter.}

\item{y}{Description of the second parameter (if applicable).}

\item{...}{Other optional parameters passed to methods.}
}
\value{
Description of the object that the function returns.
If the function doesn't return anything meaningful, you can say `NULL`.
}
\description{
Title: get_fig_eval
}
\details{
Create figures to evaluate model
}
