% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_proba.R
\name{get_proba}
\alias{get_proba}
\title{Title: get_proba}
\usage{
get_proba(MW, df_traits, plot_diag = TRUE)
}
\arguments{
\item{MW}{Adjency matrix (or metaweb)}

\item{y}{Description of the second parameter (if applicable).}

\item{...}{Other optional parameters passed to methods.}
}
\value{
Description of the object that the function returns.
If the function doesn't return anything meaningful, you can say `NULL`.
}
\description{
Title: get_proba
}
\details{
Find the best proba to validate for the adjency matrix
}
