% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traits.R
\name{infer_traits}
\alias{infer_traits}
\title{Title: infer_traits}
\usage{
infer_traits(data_traits, data_presence, method = c("mean_level", "100matrix"))
}
\arguments{
\item{x}{Description of the first parameter.}

\item{y}{Description of the second parameter (if applicable).}

\item{...}{Other optional parameters passed to methods.}
}
\value{
Description of the object that the function returns.
If the function doesn't return anything meaningful, you can say `NULL`.
}
\description{
Clean trait data. Get common lenght if missing (0.6*TL)
}
\details{
Infer trait data when missing
}
