% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{make_species_station_matrix}
\alias{make_species_station_matrix}
\title{Title: make_species_station_matrix}
\usage{
make_species_station_matrix(
  df,
  station_col = "ID.Number",
  species_col = "taxon",
  count_col = "count",
  convert_binary = TRUE
)
}
\arguments{
\item{x}{Description of the first parameter.}

\item{y}{Description of the second parameter (if applicable).}

\item{...}{Other optional parameters passed to methods.}
}
\value{
Description of the object that the function returns.
If the function doesn't return anything meaningful, you can say `NULL`.
}
\description{
Title: make_species_station_matrix
}
\details{
Transform from dna to matrix fitted for funtion of p/a, adds primary and secondary producers
}
