% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter.R
\name{metaweb_mod_parameters}
\alias{metaweb_mod_parameters}
\title{Title: metaweb_mod_parameters}
\usage{
metaweb_mod_parameters(
  data_path = NULL,
  path_par_output = NULL,
  observation_resample = 50
)
}
\arguments{
\item{x}{Description of the first parameter.}

\item{y}{Description of the second parameter (if applicable).}

\item{...}{Other optional parameters passed to methods.}
}
\value{
Description of the object that the function returns.
If the function doesn't return anything meaningful, you can say `NULL`.
}
\description{
Title: metaweb_mod_parameters
}
\details{
Find the parameter for the metaweb
}
