% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network_focus_sp}
\alias{plot_network_focus_sp}
\title{Title: plot_network_focus_sp}
\usage{
plot_network_focus_sp(MW, sp_focus, order_interaction = 1, sp_to_keep = NULL)
}
\arguments{
\item{MW}{Global metaweb}

\item{sp_focus}{Focus species where we keep them alongside all species interacting with them}

\item{order_interaction}{numeric, number of orders of interaction to keep}

\item{sp_to_keep}{Species to keep in the metaweb (species and interactions not in this vector will be removed)}
}
\value{
ggplot
}
\description{
Creates a metaweb plot, with the option of having a sub-metaweb by providing a vector and proving species focus vector to only focus on one
several species within that web
}
