% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_tree_network}
\alias{plot_tree_network}
\title{Title: plot_tree_network}
\usage{
plot_tree_network(MW, sp_to_keep = NULL)
}
\arguments{
\item{x}{Description of the first parameter.}

\item{y}{Description of the second parameter (if applicable).}

\item{...}{Other optional parameters passed to methods.}
}
\value{
Description of the object that the function returns.
If the function doesn't return anything meaningful, you can say `NULL`.
}
\description{
Create a local web wither from the global metaweb or
}
\details{
Plot network
}
