% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_tree_network}
\alias{plot_tree_network}
\title{Title: plot_tree_network}
\usage{
plot_tree_network(MW)
}
\arguments{
\item{x}{Description of the first parameter.}

\item{y}{Description of the second parameter (if applicable).}

\item{...}{Other optional parameters passed to methods.}
}
\value{
Description of the object that the function returns.
If the function doesn't return anything meaningful, you can say `NULL`.
}
\description{
Title: plot_tree_network
}
\details{
Plot network
}
